DMA: Make DMA stores visible to the rest of the system.
This required three major changes to Aladdin:

  1. Scratchpads must keep a local copy of all data that has been
     touched. This data is loaded during a dmaLoad and updated on
     stores.  This is skipped in standalone mode. During dmaStores, the
     data for each packet is no longer empty, but the actual data in the
     logical array.
  2. The TLB can now perform an INVISIBLE address translation. This
     translation takes zero time and returns the same result as the
     timing model, but it also does not update any state - no stats,
     no TLB entries insertions or evictions. This is for the DMA engine
     to use. Normally, this translation is handled by the OS so that
     everything the DMA engine sees is already physically addressed, but
     that doesn't work for our trace driven simulator, so an invisible
     translation lets us mimic this behavior without any timing
     side-effects.
  3. A dmaStore now issues a WriteInvalidate command, so that any shared
     cache lines in other caches are invalidated. This does not seem to
     have any appreciate impact on performance, which means that the
     classic cache model probably handles this in a mostly functional
     manner. This also means our current analytical flush/invalidate
     model can continue to exist (for now).

An integration test has been added to demonstrate this working.

BUG=ALADDIN-49

+++ common/ExecNode.h
@@ -28,8 +28,9 @@ struct MemAccess {
   size_t offset;
   // Is floating-point value or not.
   bool is_float;
-  // If this is not a store, then this value is meaningless.
-  double value;
+  // Hex representation of the value loaded or stored.  For FP values, this is
+  // the IEEE-754 representation.
+  uint64_t value;
 
   MemAccess() {
     vaddr = 0x0;
@@ -121,7 +122,7 @@ class ExecNode {
                       size_t offset,
                       size_t size_in_bytes,
                       bool is_float = false,
-                      double value = 0) {
+                      uint64_t value = 0) {
     mem_access = new MemAccess;
     mem_access->vaddr = vaddr;
     mem_access->offset = offset;
